requirements.txt
# EasyScale Router Agent - Python Dependencies

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

# ============================================================================
# AI/ML Stack
# ============================================================================
dspy-ai==2.5.43
langgraph==0.2.60
langchain==0.3.12
langchain-core==0.3.28

# ============================================================================
# LLM Providers
# ============================================================================
openai==1.57.2
anthropic==0.40.0
groq==0.12.0

# ============================================================================
# Database
# ============================================================================
supabase==2.10.0
psycopg[binary]==3.2.13

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.1
python-multipart==0.0.18
httpx==0.27.2

# ============================================================================
# Testing
# ============================================================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# ============================================================================
# Development
# ============================================================================
black==24.10.0
ruff==0.8.4
mypy==1.13.0

# ============================================================================
# Monitoring (Optional)
# ============================================================================
# langsmith==0.1.147
# opentelemetry-api==1.28.2
# opentelemetry-sdk==1.28.2
